Fix segfault when invoked with --debug
The short form of -d is marked as requiring an argument, while the long
option (--debug) was marked as taking an optional argument. Since the
code really doesn't handle the optional argument, we change --debug to
require an argument and update the help message accordingly.

Patch imported from Debian, author is Apollon Oikonomopoulos
<apoikos@gmail.com>